Simplify process termination slightly
This is still a mess.

+++ src/terminator/terminal/pty/PTYProcess.java
@@ -4,7 +4,6 @@ import e.util.*;
 import java.awt.Dimension;
 import java.io.*;
 import java.util.concurrent.*;
-import java.util.regex.*;
 import org.jessies.os.*;
 
 public class PTYProcess {
@@ -16,10 +15,7 @@ public class PTYProcess {
         private int pid;
         private String slavePtyName;
 
-        private boolean didDumpCore = false;
-        private boolean didExitNormally = false;
-        private boolean wasSignaled = false;
-        private int exitValue;
+        private WaitStatus status;
 
         private InputStreamReader input;
         private OutputStreamWriter output;
@@ -62,6 +58,8 @@ public class PTYProcess {
         }
 
         public void write(String string) throws IOException {
+                if (status != null)
+                        return;
                 output.write(string);
                 output.flush();
         }
@@ -72,7 +70,7 @@ public class PTYProcess {
                 invoke(new Callable<Exception>() {
                         public Exception call() {
                                 try {
-                                        waitFor0();
+                                        waitPID();
                                         return null;
                                 } catch (Exception ex) {
                                         return ex;
@@ -82,36 +80,20 @@ public class PTYProcess {
                 executorService.shutdownNow();
         }
 
-        private void waitFor0() throws IOException {
-                // FIXME: rewrite this to be more like the JDK's Process.waitFor, both in behavior and implementation.
-
-                // We now have no further use for the fd connecting us to the child, which has probably exited.
-                // Even if it hasn't, we're no longer reading its output, which may cause the child to block in the kernel,
-                // preventing it from terminating, even if root sends it SIGKILL.
-                // If we close the pipe before waiting, then we may let it finish and collect an exit status.
+        private void waitPID() throws IOException {
                 Posix.close(fd);
                 fd = -1;
 
-                // Loop until waitpid(2) returns a status or a real error.
-                WaitStatus status = new WaitStatus();
+                status = new WaitStatus();
                 int result;
-                while ((result = Posix.waitpid(pid, status, 0)) < 0) {
-                        if (result != -Errno.EINTR) {
-                                // Something really went wrong; give up.
-                                throw new IOException("waitpid(" + pid + ") failed: " + Errno.toString(-result));
-                        }
-                }
+                do {
+                        result = Posix.waitpid(pid, status, 0);
+                } while (-result == Errno.EINTR);
 
-                // Translate the status.
-                if (status.WIFEXITED()) {
-                        exitValue = status.WEXITSTATUS();
-                        didExitNormally = true;
-                }
-                if (status.WIFSIGNALED()) {
-                        exitValue = status.WTERMSIG();
-                        wasSignaled = true;
-                        didDumpCore = status.WCOREDUMP();
-                }
+                if (result < 0)
+                        throw new IOException("Waiting for PTY process failed: " +
+                                              Errno.toString(-result));
+
         }
 
         /**
@@ -130,34 +112,30 @@ public class PTYProcess {
         }
 
         public String toString() {
-                StringBuilder string = new StringBuilder();
-                string.append(String.format("PtyProcess[pid=%d,fd=%d,pty=%s]",
-                                            pid, fd, slavePtyName));
-                if (didExitNormally)
-                        string.append(",didExitNormally,exitValue=").
-                               append(exitValue);
-                if (wasSignaled)
-                        string.append(",wasSignaled,signal=").append(exitValue);
-                if (didDumpCore)
-                        string.append(",didDumpCore");
-                string.append("]");
-                return string.toString();
+                return String.format("PtyProcess[pid=%d,fd=%d,pty=%s]",
+                                     pid, fd, slavePtyName);
         }
 
         public String toExitString() {
-		if (didExitNormally)
-                        return "Process exited with status " + exitValue;
-		else if (wasSignaled)
-                        return "Process killed by " + Signal.toString(exitValue) +
-                               (didDumpCore ? " — core dumped" : "");
+                if (status == null)
+                        return "Process is still running";
+                else if (status.WIFEXITED())
+                        return "Process exited with status " + status.WEXITSTATUS();
+		else if (status.WIFSIGNALED())
+                        return "Process killed by " +
+                                Signal.toString(status.WTERMSIG()) +
+                                (status.WCOREDUMP() ? " — core dumped" : "");
 		else
                         return "Lost contact with process";
         }
 
         public void destroy() throws IOException {
+                if (status != null)
+                        return;
                 int rc = Posix.killpg(pid, Signal.SIGHUP);
                 if (rc < 0)
                         throw new IOException("Killing process failed: " +
                                               Errno.toString(-rc));
+                status = new WaitStatus();
         }
 }
